Remove nonce and todo about crypt magic
Don't care: if someone can listen to the traffic, they can also
intrude the connection even in case of tcp: spoof src IP address
is simple.
The only thing that can be done is to filter out noise from other
sources who do not know connection key, but this is not a problem for now.